requirements.txt
Flask==2.3.3
requests==2.31.0
mysql-connector-python==8.2.0
gunicorn==21.2.0